rename Login and Register page components
